veyron/runtimes/google/ipc: implement a store for the results of previous RP
This is the first in a series of CLs to implement support for allowing a
dependent RPC call to access the results of a dependee RPC call without havi
to roundtrip the results back to the client. Thus, a sequence of the form:
x := A(); B(x), can be executed as B(A()) on the server.

Change-Id: Ie69717924a1d951ae0431b78a8b7084db6d2d926