veyron/examples/rockpaperscissors: New demo game
Add a demo game of rock-paper-scissors. There are three different roles
in the game:

1. Judge: A judge enforces the rules of the game and decides who the
   winner is. At the end of the game, the judge reports the final
   score to all the score keepers.
2. Player: A player can ask a judge to start a new game, it can
   challenge another player, and it can play a game.
3. ScoreKeeper: A score keeper receives the final score for a game
   after it ended.

Each node takes the role of player, judge, and score keeper.

When the process starts, it registers itself in the mount table as
"rps/player/<hostname>", "rps/judge/<hostname>", and
"rps/scorekeeper/<hostname>".

Then each player attempts to initiate a game with another player, first
by finding a judge to run the game, and then challenging the other
player.

The players keep playing forever.

Change-Id: I5d670808b2914faf7a2fabcfc242bf5be2c4928c