veyron2/naming: rename and rationalise routines for 'terminal' names.
Replaced the 'Fixed' family of functions with ones with more meaningful
names: MakeTerminal, MakeResolveable, a test for a terminal name,
Terminal and a test for Validity. These are designed to be more composable
than the functions they replace, allbeit at the cost of in some cases asking
the caller to invoke them in combination. If these combinations turn out to be
required in user code then utility routines can easily be added.
Remove functions that are called only once (Depth) or never (StripSuffix).

Improved the documentation to better explain terminal and non-terminal
(i.e. resolveable) names.

Improved the unit tests to hopefully make them easier to comprehend.

Change-Id: Ied88d1524167b07218f962d645d0875af0f10c30