"veyron/services/identity/identity": Default Name
This CL changes the identity tool so that it generates an identity
with a default name when no flags are specified. The default name
is of the form  <user>@<hostname>, where <user> is the username of
the currently logged-in user and <hostname> is the host name as
reported by the kernel.

Change-Id: I04527cc54bda33cc512dd55ec6eaca3721547343