veyron/runtimes/google/ipc: Move "publisher" to its own package.
In unrelated design discussions, it was determined that instead of the server
publishing the proxied-endpoint, the server should instead ask the proxy to
publish the name (only the proxy will surely know its public-network
endpoint).

Move the publisher out to a separate library so that the proxy can
re-use it for publishing the names of proxied servers.

Change-Id: Iecf5d9e56a7ead82999d581d791a7e6fdd2f97a8