veyron/examples/rockpaperscissors: Fix race in test.
Make sure the test waits for all players to be idle before looking at the
counters.

In the test, there are two players. One that's controlled by the test itself,
and another one that's passively receives a challenge from the first one.
Before this change, it was possible for the first player to be done, but not
the second one at the time the counters are inspected.

Change-Id: Ibcc2a66c4f9940465201889df4e4a07717ee34db