veryon2/naming: Pipe contexts through the mounttable client library.
In the process I had to move the context into its own package.  I
think this is OK because ServerContext, of which the context should be
a subset, only depends on security currently, and security already has
the problem that it can't depend on anything.  So I think there is
actually a fair shot that we will not have to add problematic
dependencies to the context type.

I'm happy to take suggestions about the name of the type (currently
it's context.T).

Change-Id: I89529de4e881e6ee7d692796a3c37f3f339214c4