veyron/tools/naming/modules: API and modules for implementing common services.
This is used to support a very simple scripting language that
makes it easy to create and test complicated MountTable and server
configurations. Eventually, this can be made a general purpose
library and used in all of our unit tests, but it's not quite
ready for that yet, in particular it has no unittests.

Change-Id: Iffe3bb3a8384f00754758ea3d60a31cbae3f6eca