veyron/runtimes/google/lib/sshagent: Library for storing private keys in SSH agent.
This implementation is based on OpenSSH and works only on *unix systems.
Caveat with Mac OS X is that the default installation of SSH on OS X does
not support ECDSA keys. However, a fresh installation (compiled from source,
or homebrew) does support them.

The motivation for this library is to ultimately use it to store the
private key of veyron identities.

Change-Id: I26bc8dba23ba2aa35d52cabb35cdbdfa551739b9