veyron/services/store/memstore/query: Fix flaky error handling.
The TestError test was failing 1/10 times because the error
handler goroutine would get cleaned up before it had a chance
to pull the error off of the channel.  We don't need the error
channel now that evalIterator has a mutex.  It is far easier
to just use the mutex.

Also switched the abort channel to be of type struct{} to
better indicate that it is a signal-only channel.

Change-Id: I0b69eea85440be6c801f5e970001df1686e7d897