"veyron/runtimes/google/security": Mistrusted IDs should be as authorized as Unknown IDs
Andres cleverly pointed out that there is no gain in forbidding
Mistrusted PublicIDs and allowing Unknown PublicIDs -- an agent
with a Mistrusted id can always create an Unknown id (e.g., a
self-signed id) and get through.

This CL makes the following changes:
1) Modifies the implementation of the Authorize method for
chainPublicIDs to not take the trust-level of the identity
provider into account. The trust-level is instead checked
by the Names method.
2) Remove the UntrustedIDProviderPrefix constant from the
wire package and define two prefix constants for unknown
and mistrusted ids in the runtimes/security package.
3) Fix a bug in newChainPrivateID so that it returns an error
when the provided name is not a valid blessing name.

Change-Id: I29dc637dcdbbf225bcbd9411926061f79093944a