veyron/services/mounttable: Add neighborhood option.
This change adds an option to publish a mounttable server to the local
neighborhood with a given name, e.g. the hostname or something that is
expected to be unique on the neighborhood.

When this option is enabled, the local neighborhood is also mounted on
the mounttable under the "nh" prefix. This means that all the devices
should see the same thing under /<local mt>/nh.

Before this change, we could only publish endpoints to the neighborhood.
Now, we can publish any veyron address.

Change-Id: I3e3b62dcaa2b482970e92ba08d4341dcedc2ff53