{veyron,javascript}/examples/rockpaperscissors: Use unresolved names
When a player sends a challenge, it needs to include the address of the
judge that will host the game. Before this change, the judge's address
was resolved by the player initiating the game, which is fine.

However, I'm working on another change that will make it possible for different
players to resolve the same name differently depending on whether they are on
the same local network or not. So, instead of passing a resolved name, we're
now passing an unresolved one.

Change-Id: I733a9552bf67a5ff929e8e46685dbebc042995a0