veyron/examples/tunnel/tunneld: Add test.sh, fix a bug.
Add a test to verify tunneld's basic functionality. In particular, it
verifies that vsh can execute remote commands and that tunneld's object
names are present in the mounttable.

Fix a newly introduced bug where a new dispatcher was created for each
object name. The impact was that only the first Serve() call was
accepted, and only one name was published.

Change-Id: Ibe3f59280ba7aceeedb63dbaa7a1ac76e66f9f5d