veyron/services/mgmt/binary/impl: test binary server should not publish.
I investigated my suspicion from reviewing go/vcl/3050 earlier today, and
it's indeed incorrect to call Serve with a non-empty name given the
setup we have there. It happens to work because in the end we connect
directly using the endpoint, but the publish itself was failing and
spewing log errors.

Change-Id: I38592803fadd0de4dd05a81b80272c7416934d07