veyron/services/store: Iterate over duplicate paths to an object.
Previously, the store iterator pruned duplicate paths to an object. If an object
was connected to the starting node via multiple paths, the iterator included
only one (arbitrarily chosen) path. This seems unintuitive, as glob is used to
list paths, and is content-agnostic.

Glob now returns all paths that do not form cycles. Workflows expecting unique
content should use the query language.
A nice side-effect is that the size of the visited map is now
O(#out-links along the longest path).

Change-Id: Ida88f04ea30a04893c07da4d54cd8540e43d53e0