veyron/runtimes/google/rt: Avoid using global flags in the library.
The veyron/security/flag library is meant for use in "main"s.
Using it within a library can create confusion as it may not
be clear to the user what services the flags affect.

For now, use the default authorization policy in rt/mgmt.go.
The tests seem to pass with it.
If and when we need a more nuanced policy, it should be
specified as an argument or a separate flag.

TESTED=go test veyron/... veyron2/...

Change-Id: I682cfde0da55f142d691c808f183681fe08bcf02