veyron/services/store/memstore/query: Implement the "hidden" keyword.
The selection operator allows for fields to be marked as "hidden".
This makes it possible to use the field in later parts of the query
but not include the field in the result sent back to the client.

Fields are hidden by wrapping them in a special type, hiddenResult.  Just
before sending a result to the client, evalIterator looks for fields of
type hiddenResult and removes them.

Removed an invalid comment on line 111.  The type assertion is for
"chan *store.QueryResult", not "*store.QueryResult".  It isn't possible
to put a channel into the store, so we won't have the issue raised
by the comment.

Change-Id: I880c4cb08f96621996629b2377357efa91cab15d