veyron2/vdl/gen: attempt at clarifying the semantics of the stream API through comments.
As a developer, I had to either peek at veyron2/ipc/model.go, or at the
implementation in veyron/runtimes/google/ipc to figure out how to correctly use
the API generated by the stub compiler. Specifically, it wasn't clear what
happens when you call Recv, Finish, Cancel (for an incoming stream) or Send,
CloseSend, Finish, and Cancel (for an outgoing stream) in different orders on
the client.  Some things are still unclear (see TODOs) and some of my
clarifications are likely wrong.

Change-Id: I1e205cbbe271f971e43e51978fcccb390714c252