veyron/services/identity: Support for two separate OAuth client configurations.
Separate (clientid, clientsecret) pairs are needed for the web-application
(that doles out PrivateIDs) and the Veyron-Blessing service (that uses
the OAuth "installed application" flow for providing blessings).

This commit provides two flags - one for each configuration.

While at it, also converted to using html/template instead of Write-ing
the HTML and also print out the published name on the web server.

Change-Id: I4d57851350cbbd48685b74b99ec2201c4db62ef4