veyron/services/store: Don't put values of the wrong type.
Previously, if the caller tried to put a value of the wrong type into a
top-level field, the value would be placed in the implicit directory of the object,
under the same name as that field.

This fix adds a new failure class to field.Set. If the value has the wrong type,
field.Set will return SetFailedWrongType, and Put() will return an error.

Change-Id: I8a6294c5a717434909e36d51e73b82a64bcb727c