veyron/services/store/service: Remove most of the interfaces.
The existence of this directory confuses everyone.  The interfaces within
service.go potentially allow us to have multiple implementations of the store
that could all be embedded within one binary. I don't see that happening, and
duplicating the code in veyron/services/store/stored is not a big deal.
The interfaces also make it easier to embed a Store within a non-store
process (e.g. binary content server).  This is not a use-case that we want
to support anymore because it really complicates the setup with the node
manager and syncd.

This is the first in a series of changes.  I just didn't want to bite off
too much at once.

Change-Id: I370f591378d4d971397b87495f4a04e8a34c2498