veyron/examples/pipetobrowser
Add a console panic plugin to pipetobrowser.
Data piped to the browser usually terminates upon a panic.
All goroutines cease and create a lot of console spew.
This plugin wraps the goroutines into divs whose visibility can be toggled.

The console lines can be highlighted by a string value filter.
This may be adapted to a regex filter in the future.

Change-Id: Ia96f73dfe024eee738d80285f1cf78ecc6b8e037