veyron/services/identity: Support for obtaining blessing given an OAuth access token.
Motivation:
The Veyron chrome extension will use the Chrome identity API to obtain an
access token (https://developer.chrome.com/apps/identity#method-getAuthToken)
and then provide that access token to the veyron identity server to obtain
a blessing.

Nick is working on the chrome extension in https://veyron-review.googlesource.com/#/c/3580/,
while this change updates the identity service so that the access token
can be exchanged for a blessing.

Change-Id: I0cd67ed7d523e80992ff5161665211668f6f259d