"veyron/security/serialization": Serialization Library
This CL adds a general serialization library that  defines:
(1) A SigningWriter for writing data (to a given data writer) and a
signature of the data (to a given signature writer)
(2) A VerifyingReader for reading and verifying data written by a
SigningWriter.

The IDManager within wspr is modified to use this serialization library.

Change-Id: Ib685215943a089149f72133706579c4fa6cb63b6