veyron/services/wspr: HTTP handlers for adding an account and associating an account with an identity.
The /create-account route needs an OAuth2 access token.  That access
token is exchanged for a blessing from the identd server.  A new
privateID is derived from the WSPR runtime and the blessing, and that
privateID is stored in the id manager.  The name of the new privateID is
returned to the client.

The /assoc-account route is used to associate an identity with an
origin.  The client will pass in the name of the identity and the origin
they wish to associate with.  An error is returned is the identity does
not exist.

When a new web socket pipe is created, WSPR checks that it has an
identity for the web socket origin, and will return an error if it does
not.

This CL also gets rid of the {encode,decode}Identity functions that were
used previously to parse privateIDs that were sent on the wire by the
client.

Change-Id: I4efe88717369a138a4607adde8a8e0f0b17e282b