veyron/examples/bank: Comment out pbankd.
This file is commented out because we are making large changes
to the Store API and we don't want to repeatedly update this example.
Additionally, the server uses a common transaction across all requests,
so there is a pretty bad race condition that we don't want people to
copy.  The only reason we are keeping this file is because it demonstrates
how to write a custom dispatcher.

Change-Id: I9dd0c291601f96afb798edd1b00faa648811de5c