veyron/runtimes/google/ipc/stream/crypto: Import a copy of the crypto/tls package.
The authentication protocol requires TLSUnique support in the crypto/tls
package which was implemented in https://code.google.com/p/go/source/detail?r=b3759654d42d1d5ca0ac38a2d400f047ff1f7bec
(requirement for authentication protocol introduced in:
https://veyron-review.googlesource.com/#/c/3387/)

This makes the authentication protocol compilable only with a Go compiler built
from source, and eventually with Go version 1.4 (when that is released, currently
expected to be December 2014).

In order to make the authentication protocol (and thus the whole project)
buildable with release versions of the Go compiler (currently 1.3.1 is the
latest), the plan is to:

(1) Copy the crypto/tls package from Go's source tree to this source tree
(2) Setup build tags so that this copied package is used only when needed
    (i.e., when using Go prior to version 1.4)
(3) Update veyron/runtimes/google/ipc/stream/crypto/tls.go so that it
    uses this copied package for go <1.4 and the standard library package
    for go >=1.4

This commit simply introduces the copied files - adding build tags to them.
A subsequent commit will take care of (2) and (3).

The following command was used to insert the build tags:
for f in `find . -name "*.go"`; do
  sed -i -e 's|^package tls$|// +build !go1.4\
\
package tls|' $f
done

Change-Id: I33fd179b652cb561e4f18ce7a46f47a078d07378