veyron/runtimes/google/ipc: Reduce test flakiness.
Some tests were setup to always call ipc.Client.CloseSend and check
for the error. However, when there are no streaming arguments, it is possible
that the server closes the stream.Flow that is used for the RPC before the CloseSend
message is sent, and thus CloseSend will return an error.

I suspect that this "race" is the cause for much of the flakiness reported in:
https://code.google.com/p/envyor/issues/detail?id=257

Thanks to jsimsa@ for planting the idea for this race being possible
(https://veyron-review.googlesource.com/#/c/3934/)

Change-Id: I0446d48195d1bd9d8152861ab491d8c4cd40bcc4