veyron/lib/stats: New library to keep global stats.
This is an implementation of a stats repository that will be used to keep track
of all the internal counters and stats for veyron applications.

Objects in the repository are organized in a tree hierarchy and can be
looked up and updated efficiently.

Example:
  foo := stats.CreateInteger("a/b/foo")
  bar := stats.CreateFloat("a/b/bar")
  foo.Set(1)
  bar.Set(2)
The values can be retrieved with:
  vfoo := foo.Value().(int64)
  vbar := bar.Value().(float64)
  v, err := stats.Value("a/b/foo")

The respository also supports Glob().

Change-Id: I89e4e6b3312b4ab89198999eeba76b40bad02da8