veyron/tools/playground/compilerd: Support CORS requests, update Dockerfile, add README.
(Copied from https://veyron-review.googlesource.com/#/c/3978/)

This CL adds support for CORS requests, so the playground client can
talk directly to the compile backend.

It also updates the Dockerfile to work with the new "veyron" tool
version, and it adds a README about how to run the compile server
locally.

NOTE: The new playground does not actually work, because the arguments
to the identity command line tool have changed.  The call to "identity"
in builder/vbuild.go will need to be updated before this will work.

Change-Id: Ie513127b173025dd73e67f047edcb839c2747f63