veyron/services/mgmt/node/impl: fix issue with closing stdin for subcommand
Context: https://code.google.com/p/envyor/issues/detail?id=192

The root cause was a copy/paste bug: I was calling nm.CloseStdin() instead of
runNM.CloseStdin(); this was essentially a no-op.  The test was still passing
since the logic inside execScript that was trying to capture a stdin closure
and pass it along to the subcommand was also broken -- the stdin of the command
was not set, and hence the subcommand's stdin started out as closed. Hence,
the call to blackbox.WaitForEOFOnStdin() was a no-op for subcommands started
via execScript.

After fixing the nm/runNM.CloseStdin() bug, and setting stdin correctly in the
subcommand started by execScript, a bunch of code can be removed and things
actually work as designed.

Change-Id: Ib6e22b73f94d3f8e8d1db84e7cd8ea3c21463864