veyron/{services/mgmt,tools}: end-to-end tests for binaryd, buildd, and profiled
This CL adds end-to-end 'test.sh' shell tests for the binary
repository, build server, and profile repository daemons. It also
makes minor modifications to these daemons and the client tools used
for communicatining with the daemons. Further, the CL extends an
existing end-to-end test for the application repository daemon with an
additional check.

This CL depends on https://veyron-review.googlesource.com/#/c/4005/

Change-Id: I4f02e552fc4ca776f2762058f1f936eb90c77fac