veyron/tools/identity: Bugfix for some window managers.
The identity tool uses "xdg-open" on Linux to open up a browser
window. On some window managers, 'xdg-open <url>' will not return
until the browser is closed, while on others (like GNOME) it returns
right after opening <url> in a tab.

The fix is to not wait for "xdg-open" to return.

Change-Id: Ib695bec621aedb0d3b350151fdc91dc483b9cbea