veyron/services/mgmt/node/impl: factor out pieces of functionality from the unit test to make the test body smaller and easier to read and extend.
This is in preparation for implementing more app management functions (Suspend
and Resume in particular).

The only change in actual functionality is to make the app ping back to the
parent before exiting (in addition to when it comes up).  This will give us a
way to verify when the app is about to exit cleanly.

Change-Id: I8e676d86045ce85d23155abee43cb81185c3c56e